[Refactor] 카메라 기능 UI 컴포넌트 모듈화 및 구도 추천 로직 개선
- TrackingDataSource 인터페이스를 Flow 반환 방식으로 변경 및 TrackingDataSourceImpl 전면 재작성
- AIRepositoryImpl 및 CameraViewModel의 구도 추천 관련 로직(reqCompRecommend) 정리
- CameraScreenCompScreen을 CompGuideModule로 이름 변경하여 역할 명확화
- CameraScreenPreviewArea.kt 컴포넌트 자체 크기 측정 로직 제거하고 previewSize 콜백 방식으로 개선
- 바운더리 마진(20%) 추가로 조기 종료 범위 설정 기능 최적화
- CameraScreenButtons.kt를 upper, lower, common으로 분리 (버튼 기능별 컴포넌트화)
- CameraScreen 하위의 model 및 mapper 클래스(PoseUIItem, FocusRingState, ViewRate)를 각각 컴포넌트의 사용 용도에 맞게 주제별 하위 폴더(pose, focusring, viewrate)로 재배치 및 패키지 구조 갱신

TODO: 구도 트래킹 부분 안정화 작업 추가 진행 필요
