+++ presentation/src/main/kotlin/com/hanadulset/pro_poseapp/presentation/feature/camera/components/common/pose/model/PoseUIItem.kt
@@ -1,4 +1,4 @@
-package com.hanadulset.pro_poseapp.presentation.feature.camera.model
+package com.hanadulset.pro_poseapp.presentation.feature.camera.components.common.pose.model
 
 import android.net.Uri
 import android.util.SizeF
