+++ presentation/src/main/kotlin/com/hanadulset/pro_poseapp/presentation/feature/camera/components/preview/focusring/FocusRingComponent.kt
@@ -1,0 +1,45 @@
+package com.hanadulset.pro_poseapp.presentation.feature.camera.components.preview.focusring
+
+import androidx.compose.animation.core.AnimationState
+import androidx.compose.animation.core.animateTo
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.unit.dp
+
+object FocusRingComponent {
+    @Composable
+    fun FocusRing(
+        modifier: Modifier = Modifier, color: Color, pointer: Offset
+    ) {
+        val animationSize = remember {
+            AnimationState(150F)
+        }
+        val rectSize = Size(animationSize.value, animationSize.value)
+        LaunchedEffect(animationSize) {
+            animationSize.animateTo(
+                targetValue = 100F, animationSpec = tween(durationMillis = 200)
+            )
+        }
+        Canvas(
+            modifier = modifier.size(animationSize.value.dp)
+        ) {
+            drawRect(
+                color = color, topLeft = pointer.copy(
+                    pointer.x - rectSize.width / 2, pointer.y - rectSize.height / 2
+                ), size = rectSize, style = Stroke(
+                    width = 3.dp.toPx()
+                )
+            )
+            drawCircle(color, radius = 5F, center = pointer)
+        }
+    }
+}
